Implement initial keystore, crypto, and signage for consent signaling
Duchy Client
------------
- The following has been implemented
  - Functions
    - `VerifyEDPSignature`
    - `SignAndEncryptResult`
  - Use Case
    - Under test there is DuchyClientTest thats shows the possible use
case of a caller to these functions

Components
----------
`crypto` package contains the basic interfaces for crypto (encrypt/decrypt).
`crypto` also contains the subpackage `signage` for (sign and verify), and
`keystore` for storing private keys

Crypto Interface
----------------
A simple interface to encrypt and decrypt data (will eventually use Tink under
the covers)

Signage Interface
-----------------
A simple interface to sign and veriy signatures of data

Keystore
--------
This abstract class can be implemented for various types of keystorage (and
eventually KMS).  Currently there is a sample In-Memory keystore to be used
for testing and bringup.  The current design is client can store private key's
in the keystore can reteive an PrivateKeyHandle to the private key (note the object
does not actually contain the private key) that contains a refrence back to
keystore that was used.

Only the `crypto` and `signage` interfaces can access the actual underlying private
key of the PrivateKeyHandle

TODOs
-----
- Unit tests will be writen once this design approach is approved
- Tink encrypt/decrypt will be written
- "Better" KeyStore KMS implementations