Hybrid Cryptor; Data Provider and Duchy Consent (#7)
* Implement initial in memory keystore, fake crypto, and java security signage for consent signaling

Data Provider
------------
Compute participation signature

Duchy Client
------------
Verify EDP Participation Signature
Sign and Encrypt Results

Crypto Interface
----------------
A simple fake reversing interface to encrypt and decrypt data (will eventually use Tink under
the covers)

Keystore
--------
This abstract class can be implemented for various types of keystorage (and
eventually KMS).  Currently there is a sample In-Memory keystore to be used
for testing and bringup.  The current design is client can store private key's
in the keystore can retrieve an PrivateKeyHandle to the private key (note the object
does not actually contain the private key) that contains a reference back to
keystore that was used.

Co-authored-by: Kyle Johnson <kej@fb.com>
Co-authored-by: Steven Ware Jones <stevenwjones@fb.com>
Co-authored-by: Sanjay Vasandani <sanjayvas@google.com>
Co-authored-by: Steven Ware Jones <stevenwarejones@gmail.com>